refactor(agents): move runners to src/agents/runner, rename BuiltinRunner -> PiRunner
Pure file move + rename. No interface introduced, no dispatch logic
changed. Layer 0 of #616, minimal scope.

- src/agents/runners/builtin.ts -> src/agents/runner/pi-runner.ts
- src/agents/runners/claude.ts  -> src/agents/runner/claude-runner.ts
- class BuiltinRunner -> PiRunner; field/accessor renamed accordingly
